--- a/booster/library/Booster/Pattern/Implies.hs
+++ b/booster/library/Booster/Pattern/Implies.hs
@@ -126,13 +126,13 @@ runImplies def mLlvmLibrary mSMTOptions antecedent consequent =
                                 (Right simplifedSubstPatL, _) ->
                                     if patL == simplifedSubstPatL
                                         then -- we are being conservative here for now and returning "not-implied".
-                                             -- We could return implies, but the condition will contain the remainder
-                                             -- as an equality contraint, predicating the implication on that equality being true.
+                                        -- We could return implies, but the condition will contain the remainder
+                                        -- as an equality contraint, predicating the implication on that equality being true.
+
                                             doesNotImply
                                                 (sortOfPattern patL)
                                                 (externaliseExistTerm existsL patL.term)
                                                 (externaliseExistTerm existsR patR.term)
-
                                         else checkImpliesMatchTerms existsL simplifedSubstPatL existsR patR
                                 (Left err, _) ->
                                     pure . Left . RpcError.backendError $ RpcError.Aborted (Text.pack . constructorName $ err)